CIF-433 - refactor order to use the customer cart endpoint and added tests for this
