--- src/orders/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "@adobe/commerce-cif-magento-order",
-  "version": "0.0.1",
-  "description": "Adobe Commerce Integration Framework (CIF) Magento Integration Order Project",
-  "author": "Adobe Systems Inc.",
-  "license": "Apache-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/adobe/commerce-cif-magento.git"
-  },
-  "dependencies": {
-    "@adobe/commerce-cif-magento-common": "0.0.1",
-    "@adobe/commerce-cif-common": "0.1.17",
-    "@adobe/commerce-cif-model": "1.0.0"
-  },
-  "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-plugin-header": "git+https://github.com/herzog31/eslint-plugin-header.git",
-    "serverless": "1.29.0",
-    "serverless-openwhisk": "0.13.0",
-    "serverless-webpack": "4.3.0",
-    "webpack": "4.0.1"
-  },
-  "eslintConfig": {
-    "extends": "../../.eslintrc.js"
-  },
-  "scripts": {
-    "fix-serverless-permission": "chmod +x $(npm bin)/serverless",
-    "lint": "eslint -f table --ignore-path ../../.eslintignore .",
-    "deploy": "serverless deploy",
-    "remove": "serverless remove",
-    "deploy-current": "npm run deploy -- --ow-package commerce-cif-magento-order@$npm_package_version --shared",
-    "deploy-latest": "npm run deploy -- --ow-package commerce-cif-magento-order@latest --shared",
-    "deploy-suffix": "npm run deploy -- --ow-package commerce-cif-magento-order@${OW_PACKAGE_SUFFIX} --shared",
-    "deploy-package": "npm run deploy-current; npm run deploy-latest",
-    "remove-current": "npm run remove -- --ow-package commerce-cif-magento-order@$npm_package_version",
-    "remove-latest": "npm run remove -- --ow-package commerce-cif-magento-order@latest",
-    "remove-suffix": "npm run remove -- --ow-package commerce-cif-magento-order@${OW_PACKAGE_SUFFIX}",
-    "remove-package": "npm run remove-current; npm run remove-latest"
-  }
-}
