--- src/orders/serverless.yml
@@ -1,29 +1,0 @@
-service: magento
-
-package:
-  individually: true
-  exclude:
-    - ./**
-
-provider:
-  name: openwhisk
-  runtime: nodejs:default
-  ignore_certs: true
-  defaultService: magento
-  service: ${opt:service, self:provider.defaultService}
-
-custom:
-  npm-version: ${file(package.json):name}@${file(package.json):version}
-  ow-package: ${opt:ow-package, self:custom.npm-version}
-
-plugins:
-  - serverless-openwhisk
-  - serverless-webpack
-
-resources:
-  packages: ${file(../common/serverlessPackageConfig.js)}
-
-functions:
-  postOrderService:
-    name: ${self:custom.ow-package}/postOrderService
-    handler: postOrder.main
