--- src/orders/webpack.config.js
@@ -1,31 +1,0 @@
-/*******************************************************************************
- *
- *    Copyright 2018 Adobe. All rights reserved.
- *    This file is licensed to you under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License. You may obtain a copy
- *    of the License at http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software distributed under
- *    the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
- *    OF ANY KIND, either express or implied. See the License for the specific language
- *    governing permissions and limitations under the License.
- *
- ******************************************************************************/
-
-'use strict';
-
-const slsw = require('serverless-webpack');
-const path = require('path');
-
-module.exports = {
-    entry: slsw.lib.entries,
-    // devtool: 'source-map',
-    target: 'node',
-    stats: 'minimal',
-    output: {
-        libraryTarget: 'commonjs',
-        path: path.resolve(__dirname, '.webpack'),
-        filename: '[name].js'
-    }
-};
-
