+++ test/carts/postOrderTest.js
@@ -18,6 +18,7 @@ const assert = require('chai').assert;
 const setup = require('../lib/setupTest').setup;
 const config = require('../lib/config').config;
 const requestConfig = require('../lib/config').requestConfig;
+const specsBuilder = require('../lib/config').specsBuilder;
 
 describe('Magento postOrder', () => {
     describe('Unit tests', () => {
@@ -25,28 +26,27 @@ describe('Magento postOrder', () => {
         // Add helpers to context
         setup(this, __dirname, 'postOrder');
 
-        it('returns an error for a missing expected parameter', () => {
+        it('returns an error for missing cartId parameter', () => {
             return this.prepareReject(null).execute(null).then(result => {
                 assert.strictEqual(result.response.error.name, 'MissingPropertyError');
             });
         });
 
-        it('returns an order', () => {
-            let args = {
-                cartId: 'dummy-id'
-            };
-            return this.prepareResolve('"12"', requestConfig(encodeURI(`http://${config.MAGENTO_HOST}/rest/V1/guest-carts/${args.cartId}/order`), 'PUT'))
-                .execute(Object.assign(args, config))
-                .then(result => {
-                    assert.isDefined(result.response);
-                    assert.strictEqual(result.response.statusCode, 201);
-                    assert.isDefined(result.response.headers);
-                    assert.strictEqual(result.response.headers.Location, 'orders/12');
-                    assert.isDefined(result.response.body);
-                    assert.isObject(result.response.body);
-                    assert.strictEqual(result.response.body.id, 12);
-                });
-
+        specsBuilder().forEach(spec => {
+            it(`returns an order for a ${spec.name} cart`, () => {
+                return this.prepareResolve('12', requestConfig(encodeURI(`http://${config.MAGENTO_HOST}/rest/V1/${spec.baseEndpoint}/order`),
+                    'PUT', spec.token))
+                    .execute(Object.assign({cartId: spec.args.id, __ow_headers: spec.args.__ow_headers}, config))
+                    .then(result => {
+                        assert.isDefined(result.response);
+                        assert.strictEqual(result.response.statusCode, 201);
+                        assert.isDefined(result.response.headers);
+                        assert.strictEqual(result.response.headers.Location, 'orders/12');
+                        assert.isDefined(result.response.body);
+                        assert.isObject(result.response.body);
+                        assert.strictEqual(result.response.body.id, 12);
+                    });
+            });
         });
 
     });
